url_shortener/requirements.txt
```
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
geoip2==4.7.0
requests==2.31.0
```